refactor: Extract helper to utils
To be able to call helper from script without importing all mjml libs
