+++ src/ducks/notifications/BalanceLower/index.js
@@ -6,7 +6,7 @@ import { getCurrencySymbol } from 'utils/currencySymbol'
 import { getCurrentDate } from 'ducks/notifications/utils'
 import template from './template.hbs'
 import { toText } from 'cozy-notifications'
-import { ACCOUNT_DOCTYPE, GROUP_DOCTYPE } from 'doctypes'
+import { ruleAccountFilter } from 'ducks/settings/ruleUtils'
 
 const addCurrency = o => ({ ...o, currency: '€' })
 
@@ -49,26 +49,6 @@ const customToText = cozyHTMLEmail => {
   return toText(cozyHTMLEmail, getContent)
 }
 
-const doesAccountCorrespondToAccountGroup = (
-  groups,
-  accountOrGroup
-) => account => {
-  if (!accountOrGroup) {
-    return true
-  } else if (accountOrGroup._type === ACCOUNT_DOCTYPE) {
-    return account._id === accountOrGroup._id
-  } else if (accountOrGroup._type === GROUP_DOCTYPE) {
-    const group = groups.find(group => accountOrGroup._id === group._id)
-    if (group && group.accounts) {
-      return group.accounts.includes(account._id)
-    } else {
-      // In case of non existent group, prefer to consider that no accounts
-      // belong to it
-      return false
-    }
-  }
-}
-
 class BalanceLower extends NotificationView {
   constructor(config) {
     super(config)
@@ -78,10 +58,8 @@ class BalanceLower extends NotificationView {
 
   filterForRule(rule, account) {
     const isBalanceUnder = getAccountBalance(account) < rule.value
-    const correspondsAccountToGroup = doesAccountCorrespondToAccountGroup(
-      this.data.groups,
-      rule.accountOrGroup
-    )(account)
+    const accountFilter = ruleAccountFilter(rule, this.data.groups)
+    const correspondsAccountToGroup = accountFilter(account)
     const isNotCreditCard = account.type !== 'CreditCard'
     return isBalanceUnder && correspondsAccountToGroup && isNotCreditCard // CreditCard are always in negative balance
   }
